feat: 문제를 풀면 quiz_score에 점수가 추가되게 구현

+++ src/main/java/soongsil/kidbean/server/summary/domain/QuizScore.java
@@ -64,7 +64,7 @@ public class QuizScore {
         totalScore = totalScore - Level.getPoint(beforeLevel) + Level.getPoint(afterLevel);
     }
 
-    public QuizScore addScore(int score) {
+    public QuizScore addScore(Long score) {
         totalScore = totalScore + score;
         return this;
     }
